debian: add python3-typing-extensions
why: fastapi and its dependencies have been backported from bullseye,
which run with python3.9. To support new typing with older python
version (3.7), we need to install a compatibility package
"python3-typing-extensions"
This package can be removed when we will be on a newer python version
(i.e. when we will migrate to bullseye)