remove uvloop dependency
why: after tests with websocketd, uvloop is not much better than the
standard implementation. Moreover, the uvloop from bullseye is not fully
compatible with python3.9 ... It doesn't cause any issue for sysconfd,
but have an issue for websocketd. This service is also not intensively
used. For these reasons and reduce dependencies, uvloop could be
removed.